Refactor Storage into suitable subclasses
